mcpServers/voice_server/requirements.txt
fastapi
uvicorn[standard]
python-multipart
coqui-tts
faster-whisper
torch
torchaudio
pydub
llama-cpp-python 
